Well unfortunaly it looks like the brute-force subtraction method isn't going to work. There is the possiblity of exploring some different angles to attack, but I feel as though the general concept is like trying to paint a skyscraper by throwing paint balloons a gajillion times. It either takes too long, or you're going to have to cheat and then it's not going to compress anything.
In fact, using this compression method, the file is actually larger because of the use of the UTF-8/ASCII encoding to save the String
output of the encoding algorithm. So, I've given up on that aspect. There may be a way to find as close to the number as possible, but
it seems like that would take a large amount of time as well.

I have started working on using division in order to make the process much shorter. This may work better but there's still a lot of
experimenting left to do. This will hopefully speed up the process significantly by using divisors. Right now, I have to debug why
the length of my big int isn't getting any smaller. Seems like it is, but there's always more work to be done.